+++ src/Config/AutoloadConfigInterface.php
@@ -32,4 +32,6 @@ interface AutoloadConfigInterface
     public function getNamespacePrefix(): ?string;
 
     public function getPackagesToCopy(): array;
+
+    public function getPackagesToPrefix(): array;
 }