+++ src/Pipeline/ChangeEnumerator.php
@@ -8,66 +8,51 @@
 namespace BrianHenryIE\Strauss\Pipeline;
 
 use BrianHenryIE\Strauss\Config\ChangeEnumeratorConfigInterface;
-use BrianHenryIE\Strauss\Files\DiscoveredFiles;
-use BrianHenryIE\Strauss\Files\FileWithDependency;
+use BrianHenryIE\Strauss\Files\File;
 use BrianHenryIE\Strauss\Types\ClassSymbol;
+use BrianHenryIE\Strauss\Types\DiscoveredSymbol;
 use BrianHenryIE\Strauss\Types\DiscoveredSymbols;
-use BrianHenryIE\Strauss\Types\FunctionSymbol;
 use BrianHenryIE\Strauss\Types\NamespaceSymbol;
-use League\Flysystem\FilesystemReader;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LoggerInterface;
-use Psr\Log\NullLogger;
 
 class ChangeEnumerator
 {
     use LoggerAwareTrait;
 
     protected ChangeEnumeratorConfigInterface $config;
-    protected FilesystemReader $filesystem;
 
     public function __construct(
         ChangeEnumeratorConfigInterface $config,
-        FilesystemReader $filesystem,
-        ?LoggerInterface $logger = null
+        LoggerInterface $logger
     ) {
         $this->config = $config;
-        $this->filesystem = $filesystem;
-        $this->setLogger($logger ?? new NullLogger());
-    }
-
-    public function markFilesForExclusion(DiscoveredFiles $files)
-    {
-
-        foreach ($files->getFiles() as $file) {
-            if ($file instanceof FileWithDependency) {
-                if (in_array(
-                    $file->getDependency()->getPackageName(),
-                    $this->config->getExcludePackagesFromPrefixing(),
-                    true
-                )) {
-                    $file->setDoPrefix(false);
-                    continue;
-                }
-
-                foreach ($this->config->getExcludeFilePatternsFromPrefixing() as $excludeFilePattern) {
-                    // TODO: This source relative path should be from the vendor dir.
-                    // TODO: Should the target path be used here?
-                    if (1 === preg_match($excludeFilePattern, $file->getSourcePath())) {
-                        $file->setDoPrefix(false);
-                    }
-                }
-            }
-        }
+        $this->setLogger($logger);
     }
 
     public function determineReplacements(DiscoveredSymbols $discoveredSymbols): void
     {
         $discoveredNamespaces = $discoveredSymbols->getDiscoveredNamespaces();
 
+        $this->determineExclusions($discoveredSymbols);
+
         foreach ($discoveredNamespaces as $symbol) {
+            // This line seems redundant.
             if ($symbol instanceof NamespaceSymbol) {
                 $namespaceReplacementPatterns = $this->config->getNamespaceReplacementPatterns();
+
+                if (in_array(
+                    $symbol->getOriginalSymbol(),
+                    $this->config->getExcludeNamespacesFromPrefixing(),
+                    true
+                )) {
+                    $symbol->setDoRename(false);
+                }
+
+                // If any of the files the symbol was found in are marked not to prefix, don't prefix the symbol.
+                if ($symbol->getNamespace() !== '\\' && !array_reduce($symbol->getSourceFiles(), fn(bool $carry, File $file) => $carry && $file->isDoPrefix(), true)) {
+                    $symbol->setDoRename(false);
+                }
 
                 // `namespace_prefix` is just a shorthand for a replacement pattern that applies to all namespaces.
 
@@ -119,13 +104,13 @@ class ChangeEnumerator
             $discoveredSymbols->getAllClasses()
         );
 
-        foreach ($classesTraitsInterfaces as $theclass) {
-            if (str_starts_with($theclass->getOriginalSymbol(), $classmapPrefix)) {
+        foreach ($classesTraitsInterfaces as $symbol) {
+            if (str_starts_with($symbol->getOriginalSymbol(), $classmapPrefix)) {
                 // Already prefixed / second scan.
                 continue;
             }
 
-            if ($theclass->getNamespace() === '\\') {
+            if ($symbol->getNamespace() === '\\') {
                 if ($symbol instanceof ClassSymbol) {
                     // Don't double-prefix classnames.
                     if (str_starts_with($symbol->getOriginalSymbol(), $this->config->getClassmapPrefix())) {
@@ -136,27 +121,26 @@ class ChangeEnumerator
                 }
             }
 
-            if ($theclass->getNamespace() !== '\\') {
-                $newNamespace = $discoveredNamespaces[$theclass->getNamespace()];
+            // If we're a namespaced class, apply the fqdnchange.
+            if ($symbol->getNamespace() !== '\\') {
+                $newNamespace = $discoveredNamespaces[$symbol->getNamespace()];
                 if ($newNamespace) {
-                    // TODO: This should be in ChangeEnumerator.
-                    // `str_replace` was replacing multiple. This stops after one. Maybe should be tied to start of string.
-                    $determineReplacement = function ($originalNamespace, $newNamespace, $fqdnClassname) {
-                        $search = '/'.preg_quote($originalNamespace, '/').'/';
-                        return preg_replace($search, $newNamespace, $fqdnClassname, 1);
-                    };
-                    $theclass->setReplacement(
-                        $determineReplacement(
-                            $newNamespace->getOriginalSymbol(),
-                            $newNamespace->getReplacement(),
-                            $theclass->getOriginalSymbol()
-                        )
+                    $replacement = $this->determineNamespaceReplacement(
+                        $newNamespace->getOriginalSymbol(),
+                        $newNamespace->getReplacement(),
+                        $symbol->getOriginalSymbol()
                     );
-                    unset($newNamespace);
+
+                    $symbol->setReplacement($replacement);
+
+                    unset($newNamespace, $replacement);
                 }
                 continue;
+            } else {
+                // Global class.
+                $replacement = $classmapPrefix . $symbol->getOriginalSymbol();
+                $symbol->setReplacement($replacement);
             }
-            $theclass->setReplacement($classmapPrefix . $theclass->getOriginalSymbol());
         }
 
         $functionsSymbols = $discoveredSymbols->getDiscoveredFunctions();
@@ -173,5 +157,42 @@ class ChangeEnumerator
 
             $symbol->setReplacement($functionPrefix . $symbol->getOriginalSymbol());
         }
+    }
+
+    protected function determineExclusions(DiscoveredSymbols $discoveredSymbols)
+    {
+        foreach ($discoveredSymbols->getSymbols() as $symbol) {
+            $this->checkExcludedSymbol($symbol);
+        }
+    }
+
+    protected function checkExcludedSymbol(DiscoveredSymbol $symbol): void
+    {
+        foreach ($this->config->getExcludeNamespacesFromPrefixing() as $excludeNamespace) {
+            $excludeNamespace = rtrim($excludeNamespace, '\\');
+            if (str_starts_with($symbol->getNamespace(), $excludeNamespace)) {
+                $symbol->setDoRename(false);
+            }
+        }
+        /** @var File $file */
+        foreach ($symbol->getSourceFiles() as $file) {
+            $relativePath = $file->getVendorRelativePath();
+            foreach ($this->config->getExcludeFilePatternsFromPrefixing() as $filePattern) {
+                if (1 === preg_match($filePattern, $relativePath)) {
+                    $symbol->setDoRename(false);
+                }
+            }
+        }
+    }
+
+
+    /**
+     *`str_replace` was replacing multiple. This stops after one. Maybe should be tied to start of string.
+     */
+    protected function determineNamespaceReplacement(string $originalNamespace, string $newNamespace, string $fqdnClassname): string
+    {
+            $search = '/' . preg_quote($originalNamespace, '/') . '/';
+
+            return preg_replace($search, $newNamespace, $fqdnClassname, 1);
     }
 }