+++ src/Pipeline/FileCopyScanner.php
@@ -16,14 +16,15 @@
 
 namespace BrianHenryIE\Strauss\Pipeline;
 
+use BrianHenryIE\Strauss\Composer\ComposerPackage;
 use BrianHenryIE\Strauss\Config\FileCopyScannerConfigInterface;
 use BrianHenryIE\Strauss\Files\DiscoveredFiles;
+use BrianHenryIE\Strauss\Files\File;
 use BrianHenryIE\Strauss\Files\FileBase;
 use BrianHenryIE\Strauss\Files\FileWithDependency;
 use BrianHenryIE\Strauss\Helpers\FileSystem;
 use BrianHenryIE\Strauss\Types\DiscoveredSymbol;
 use BrianHenryIE\Strauss\Types\NamespaceSymbol;
-use League\Flysystem\FilesystemReader;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
@@ -34,7 +35,7 @@ class FileCopyScanner
 
     protected FileCopyScannerConfigInterface $config;
 
-    protected FileSystem $fileSystem;
+    protected FileSystem $filesystem;
 
     public function __construct(
         FileCopyScannerConfigInterface $config,
@@ -42,7 +43,7 @@ class FileCopyScanner
         ?LoggerInterface $logger = null
     ) {
         $this->config = $config;
-        $this->fileSystem = $filesystem;
+        $this->filesystem = $filesystem;
 
         $this->setLogger($logger ?? new NullLogger());
     }
@@ -53,41 +54,30 @@ class FileCopyScanner
         foreach ($files->getFiles() as $file) {
             $copy = true;
 
-            if ($file instanceof FileWithDependency) {
-                if (in_array($file->getDependency()->getPackageName(), $this->config->getExcludePackagesFromCopy(), true)) {
-                    $this->logger->debug("File {$file->getSourcePath()} will not be copied because {$file->getDependency()->getPackageName()} is excluded from copy.");
-                    $copy = false;
-                }
-            }
-
             if ($this->config->getTargetDirectory() === $this->config->getVendorDirectory()) {
                 $this->logger->debug("The target directory is the same as the vendor directory."); // TODO: surely this should be outside the loop/class.
                 $copy = false;
             }
 
-            /** @var DiscoveredSymbol $symbol */
-            foreach ($file->getDiscoveredSymbols() as $symbol) {
-                foreach ($this->config->getExcludeNamespacesFromCopy() as $namespace) {
-                    if (in_array($file->getSourcePath(), array_keys($symbol->getSourceFiles()), true)
-                        && $symbol instanceof NamespaceSymbol
-                        && str_starts_with($symbol->getOriginalSymbol(), $namespace)
-                    ) {
-                        $this->logger->debug("File {$file->getSourcePath()} will not be copied because namespace {$namespace} is excluded from copy.");
-                        $copy = false;
-                    }
+            if ($file instanceof FileWithDependency) {
+                if ($this->isPackageExcluded($file->getDependency())) {
+                    $copy = false;
+                    $this->logger->debug("File {$file->getSourcePath()} will not be copied because {$file->getDependency()->getPackageName()} is excluded from copy.");
                 }
             }
 
-            $filePath = $file->getSourcePath();
-            foreach ($this->config->getExcludeFilePatternsFromCopy() as $pattern) {
-                if (1 == preg_match($pattern, $filePath)) {
-                    $this->logger->debug("File {$file->getSourcePath()} will not be copied because it matches pattern {$pattern}.");
-                    $copy = false;
-                }
+            if ($this->isNamespaceExcluded($file)) {
+                $copy = false;
+            }
+
+            if ($this->isFilePathExcluded($file->getSourcePath())) {
+                $copy = false;
             }
 
             if ($copy) {
-                $this->logger->debug("Marking file {$file->getSourcePath()} to be copied.");
+//                $this->logger->debug("Marking file {relativeFilePath} to be copied.", [
+//                    'relativeFilePath' => $this->filesystem->getRelativePath($this->config->getVendorDirectory(), $file->getSourcePath()),
+//                ]);
             }
 
             $file->setDoCopy($copy);
@@ -98,8 +88,80 @@ class FileCopyScanner
 
             $file->setAbsoluteTargetPath($target);
 
-            $shouldDelete = $this->config->isDeleteVendorFiles() && ! $this->fileSystem->isSymlinkedFile($file);
+            $shouldDelete = $this->config->isDeleteVendorFiles() && ! $this->filesystem->isSymlinkedFile($file);
             $file->setDoDelete($shouldDelete);
+
+            // If a file isn't copied, don't unintentionally edit the source file.
+            if (!$file->isDoCopy() && $this->config->getTargetDirectory() !== $this->config->getVendorDirectory()) {
+                $file->setDoPrefix(false);
+            }
+            if (!$copy && $this->config->getTargetDirectory() !== $this->config->getVendorDirectory()) {
+                foreach ($file->getDiscoveredSymbols() as $symbol) {
+                    $symbol->setDoRename(false);
+                }
+            }
         };
+    }
+
+    protected function isPackageExcluded(ComposerPackage $package): bool
+    {
+        if (in_array(
+            $package->getPackageName(),
+            $this->config->getExcludePackagesFromCopy(),
+            true
+        )) {
+            return true;
+        }
+        return false;
+    }
+
+    protected function isNamespaceExcluded(File $file): bool
+    {
+        /** @var DiscoveredSymbol $symbol */
+        foreach ($file->getDiscoveredSymbols() as $symbol) {
+            if (!($symbol instanceof NamespaceSymbol)) {
+                continue;
+            }
+            foreach ($this->config->getExcludeNamespacesFromCopy() as $namespace) {
+                $namespace = rtrim($namespace, '\\');
+                if (in_array($file->getSourcePath(), array_keys($symbol->getSourceFiles()), true)
+                    // TODO: case insensitive check. People might write BrianHenryIE\API instead of BrianHenryIE\Api.
+                    && str_starts_with($symbol->getOriginalSymbol(), $namespace)
+                ) {
+                    $this->logger->debug("File {$file->getSourcePath()} will not be copied because namespace {$namespace} is excluded from copy.");
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Compares the relative path from the vendor dir with `exclude_file_patterns` config.
+     *
+     * @param string $absoluteFilePath
+     * @return bool
+     */
+    protected function isFilePathExcluded(string $absoluteFilePath): bool
+    {
+        foreach ($this->config->getExcludeFilePatternsFromCopy() as $pattern) {
+            $escapedPattern = $this->preparePattern($pattern);
+            if (1 === preg_match($escapedPattern, $absoluteFilePath)) {
+                $this->logger->debug("File {$absoluteFilePath} will not be copied because it matches pattern {$pattern}.");
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private function preparePattern(string $pattern): string
+    {
+        $delimiter = '#';
+
+        if (substr($pattern, 0, 1) !== substr($pattern, - 1, 1)) {
+            $pattern = $delimiter . $pattern . $delimiter;
+        }
+
+        return $pattern;
     }
 }